+++ mrb_doc/core/cursor.rb
@@ -1,27 +1,3 @@
-# Shows the cursor.
-# @return [nil]
-def show_cursor
-  # mrb_show_cursor
-  # src/mruby_integration/core/cursor.cpp
-  nil
-end
-
-# Hides the cursor.
-# @return [nil]
-def hide_cursor
-  # mrb_hide_cursor
-  # src/mruby_integration/core/cursor.cpp
-  nil
-end
-
-# Checks if the cursor is hidden.
-# @return [Boolean]
-def cursor_hidden?
-  # mrb_cursor_hidden
-  # src/mruby_integration/core/cursor.cpp
-  true
-end
-
 # Enables and shows the cursor.
 # @return [nil]
 def enable_cursor